exceptions: Added LoadErrorReason
This is used especially to test the various ways in which a load can
fail, we keep these under the same LoadError exception because you can't
really recover from any of these errors and it's more practical to handle
as one exception. The reason only tells you a bit about why the exception
was raised, and it's better to test this than to do string compares in
test cases.