buildelement.py: Use a sorted dict and not an ordered one.
This change causes cache keys to be stable (without this
they are different every time due to some random order).

The input of the element does not change if you specify the
install commands before the build commands, and in any case
I dont think we can rely on knowing the order dict members
were specified in a yaml file.