context.py: Now stores a frontent message handler.
All messages are proxied through the context which is shared
across pipelines and created by the frontent. The frontend
can override how logging is done.

In sub processes, the frontend of the element is the scheduler
job itself, which then proxies messages through a Queue object
(implented with a pipe).