_ostree.py: Adding ostree module for handling ostree interactions.
We share some code this way and also have the option of replacing
the implementation for a shell based one instead of a gi bindings
based implementation (which would allow us to depend on older versions
of ostree).

Note we use repo.remote_change() API instead of repo.remote_add(),
this allows to use a create-if-not-exists flag for configuring
remotes.