element.py: Messing with caches
Caches are not working, trying to force them for now.

Ultimately we need to snapshot cached stated of elements once
at startup and then update the local data model based on successful
completion of tasks in child processes. Introspecting cached
state is expensive as it involves interrogating the artifact cache
and plugins which may run host tools or shell commands.