element.py: Allow variable substitutions in environment variables.
There is some danger in confusing the environment variables with
system targetting variables, one should not make assumptions around
the install %{prefix} being the correct prefix for running, even
though it's correct some of the time and tempting to do.

On the other hand variable substitutions here can be useful.