element.py: Include build dependency integration commands in cache keys.
Integration commands in BuildStream run whenever staging the build
dependencies of an element, so that an element can expect a fully
integrated base to build on.

As such, integration commands imposed on depending elements do not
effect the given element's cache key. However, the sum of an element's
dependency integration commands does effect the depending element's
cache key.