_loader.py: Fixed expensive circular dependency checks
The last circular dependency checker was insanely convoluted and
also insanely expensive.

The dependency check is still performed at the correct place: after
resolving the variants and knowing what the dependencies will be
for the pipeline, and just before returning LoadElements.

Dependency check is now also performed in linear time.