_main.py: Print some feedback while loading the pipeline.
Enourmous pipelines which have around 450 elements, such as
a full stack OS with GNOME components, will take around 5 seconds
to load, better to show the user what's going on at load time.

This has been optimized as much as possible within reason, we need
to interrogate the artifact cache (io) and check source consistency
(with git this is done with git cat-file to check if the ref is present
in a local cache), these parts just take time.