_sandboxbwrap.py: Force terminate sandboxed jobs when terminating a task
This is a tricky problem because bwrap is a setuid root program which
we do not have permission to terminate.

To solve this, we ensure --unshare-pid is used (which should have been
there already probably anyway) which causes bwrap to create one additional
direct child process owning the created pid namespace.

Using psutil, we enumerate the direct child of the spawned bwrap process,
which we do have permission to kill, and kill that task directly.

Since all child tasks are spawned in a pid namespace, they are automatically
killed once the owning process of that namespace exits.