_scheduler.py: Added ability to terminate ongoing tasks and SIGINT handling
The scheduler now handles SIGINT in the main loop while processing and
allows the caller to make a decision of what to do when ^C is pressed.

SIGINT is blocked for child jobs, and when the new terminate_jobs()
is invoked, then child tasks will be sent SIGTERM, which they will
use to exit as gracefully as possible by flushing any open log files
to disk and cleaning up any runtime state.