_pipeline.py: Some changes of how the pipeline api works.
The pipeline no longer owns the scheduler but is given a scheduler
to operate on, this needs to be pushed a little further for recursive
pipelines to work with a single application scheduler.

Also consequently, this patch addresses a problem in pipeline.fetch()
which was failing to filter down to only the elements which needed
fetching (was resulting in noop of fetch when actually some elements
required a fetch).