utils.py: Added _suspendable() context manager.
This takes care of handling SIGTSTP and SIGCONT, allowing the
caller to pass one suspend() and one resume() function. This
should be used in the core where we launch sub processes which
will not go to sleep with the scheduler job, so that the handlers
in place can manually put these child processes to sleep and
wake them up later.