_scheduler.py: Handle SIGTSTP in main process.
This is damn tricky because we also use it to put child processes to sleep.

However the result of this, is that now the application behaves
as a user would expect when ^Z is issued in the calling terminal,
whether or not the scheduler currently has jobs suspended, the
result is correct.