_main.py: Handle new scheduler signals and print status info
Also now added --no-interactive option, resolve interactive mode
by first checking if interaction was explicitly disabled, and then
checking if we are connected to a proper terminal or not.

Currently we do not distinguish between being connected to a
terminal and whether the terminal supports styling, which is
almost always the same, but the more correct check would be
just to check if the input stream isatty().