main.py: Use is_a_tty property instead of does_styling to detect interactive mode.
The state of things is a bit of a mess, in fact both click and blessings
detect this with either a force override and falling back to isatty() directly
to determine whether we are connected to a terminal which supports the
ANSI color codes we desire.

More complex libraries like ncurses, maintain their own termcap database
to decide, if they are connected to a terminal, whether that particular
terminal supports specific ANSI escape codes, which is also pretty overkill.