element.py: Added stage_dependencies() convenience function.
This iterates over the elements dependencies and stages them somewhere
in a sandbox.

This is especially interesting because of the added warnings it gives
you about the file overlaps and ignored files which would otherwise
overwrite non-empty directories.

Also updated build and compose elements to use this in place of
manually looping.