_ostree.py: Changes to repository modes.
We use ostree primarily for the artifact cache and also for
the ostree source plugin, for the artifact cache, we want "bare user"
mode repository.

Changes are that:
  o We no longer encode files as uid/gid root, because it requires "user mode"
    option at checkout time
  o We no longer use "user mode" option when checking out the artifacts (but
    we still need to use it for the ostree source).

This is all because checking out with "user mode" not only drops the ownership
bits on the checked out files, but it _also_ munges the permission bits, we
need the file permission bits to be preserved.