git.py source plugin: Fixed staging of submodules.
This was failing only the first time around, if the pipeline was
loaded and the git mirror for the primary module is not yet present.

Need to re-introspect the presence of submodules at stage time
just so that the GitSource knows to stage submodules which is previously
fetched at fetch() time.