_signals.py: More reliable suspend/resume context manager
It seems that resuming tasks was not happening consistently, sometimes
resuming would fail to resume.

Changed the code to stop handling SIGCONT completely as to not add
any interference in the main child task's SIGCONT handling, instead
perform SIGCONT activities in the SIGTSTP handler, directly after
issuing SIGSTOP to self, where we implicitly know that if that code
is executed, it's because we have now resumed.

It's impossible to tell how correct this change is, however after
much testing this appears to be entirely reliable, whilst the previous
approach continues to be demonstrably unreliable.