_frontend/main.py: Fault tollerance in handling failures
First, dont offer the option to shell into a failed build sandbox
if there is no sandbox, this happens when it is a fetch operation
that fails.

Second, dont continue to handle further failures when the user
has opted to terminate; this can happen when more than one task
simultaneously fails.