status.py and main.py: Added StatusHeader widget to status area
Now the Status() widget is given a reference to the pipeline and
scheduler and the status area fetches the elapsed time by itself
without needing to be driven by the main.py frontend.

The new StatusHeader area displays:
  o Main invocation target
  o Overall elapsed time
  o Total elements in the pipeline to be processed
  o Number of elements which need processing in the session
  o For each Scheduler Queue, the number of elements already
    processed by that queue

Also, now the status widgets rely on an external elapsed time
offset which is provided by the scheduler instead of calculating
time locally, just use the scheduler elapsed time offset.