_sandboxbwrap.py: Use hard coded list of device nodes to mount
It does not make much sense to have devices as a project configuration,
especially if we want to eventually have a windows port, we would rather
keep sandbox devices as a constant and internal detail of the sandbox
implementation.

Note that this list intentionally omits /dev/console which was previously
part of the devices exposed to the sandbox configured by the project,
however the console device is not required to build and not present
on some auto builders without a console.