Documentation: Added user facing project configuration section
Instead of documenting this in the Project object, provide a section
more targetted at users. The default configuration is shown in the
user facing documentation and removed from the Project object documentation
which is more targetted at API references for plugin authors.

References to the Project Configuration from the format documentation
have been updated to refer to this new section.