Added __init__.py to the plugin subdirectories.
This makes the 'sources' and 'elements' subdirectores modules technically,
but it does not effect how we load them with pluginbase, that still works.

Updated documentation machinery to have buildstream/plugins in the PYTHONPATH
and import the docs as elements.autotools etc.

This is all because since recent sphinx started importing from distutils,
this was conflicting with our distutils plugin.