Note that OSTree v2017.6 or newer is preferred
There is a GPG bug that can cause a 10 second delay when loading
pipelines, if the pipeline includes a signed OSTree repo and the
path to the repository is over a certain length.

OSTree v2017.6 and newer have internal changes that make hitting this
issue much less likely, so let's make users aware of that.