_pipeline.py: Rewrite elements during track, instead of at the end.
This was originally written to be at the end because we thought it
may be possible to have multiple tasks which effect the same file,
but this is now untrue since we removed the include feature and
the scheduler always runs things at element granularity (one element
one file, one element processed at a time).

Since it's more desirable to rewrite a tracked source immediately
after tracking it, that's what we're doing now.