queue.py: Check whether to skip elements every time we reschedule
Instead of only checking which elements to skip at queue insertion
time, check whether to skip elements again directly before running
them when they are in a ready state.

This means the decision to skip can be made at any time before the
element is ready, and is important in scenarios with track queues
in play where cache keys can change (or be assigned) along the way.