element.py: Added _force_inconsistent() method and fixed cached state
For initializing a pipeline in a forcefully inconsistent state,
this is required for pipelines which execute TrackQueues.

Also improved handling of the cached state in general so that the cached
state is only calculated once, and only after there is a cache key. The
element appears to be not cached unconditionally until a cache key
can be resolved.