_sandboxbwrap.py: Cleanup debris left behind by bwrap
When running a script element which intends to collect the
entire mutated sysroot as an output artifact, it is important to
clean up any debris which bwrap created.

This includes the tmpfs dir it creates automatically and the
device files we bind mount into the sandbox, we especially dont
want these around because they may have real root gid ownership
due to bubblewrap being a setuid root binary, and collecting these
into an artifact will cause problems later on.