Adding _fuse submodule for fuse layer implementations.
This brings in a base Mount class which takes care of synchronizing
a mount in a sub process (fork()).

Additionally, a SafeHardlinks Mount implementation was added which
implements a copy-on-write hardlink fuse layer.

This ensures that any time a file is opened for writing, the underlying
file is copied in the case that it is a hardlink (these copies are not
applied in the cases where proper 'install' semantics are applied and
the file is removed first).