element.py: Split up assemble into three methods.
Some elements may continue to do everything inside assemble(),
however they will not support shelling in properly, this is sometimes
normal, as with import elements.

The shell implementation has been enhanced with this, as it will
now setup the shell in Scope.BUILD scope the same way the element
would when it's about to assemble.