main.py: Few changes with rendering status and job termination
Created a maybe_render_status() helper function instead of having
conditionals all over the place.

Always resume jobs in the interrupted() context manager, even
if we're going to terminate, also just call the maybe_render_status()
there unconditionally so we dont care if we're terminating or not
in that context manager.