WIP: add `bst source-bundle` command
This command will generate a tarball containing the sources and build
instructions needed to produce an element. The build can then be
manually transferred to another host and executed.

This means that the resulting build lacks the usual reproducibility
guarantees that BuildStream would provide. This command should only
be used for bootstrapping new architectures where it's not yet
possible to run a normal BuildStream build.

We hope to obsolete this command in future by making smarter Sandbox
implementations which can allow BuildStream to drive the whole
process, using QEMU to avoid the need for real hardware until we
can produce a clean, deployable image.