import.py: Raise an error for elements that have no sources
Previously the build would start but then fail during assemble() once it
became clear there were no files to import. Now an error is raised right
away:

        Error loading pipeline: import element at gnu-toolchain/base-platform.bst [line 1 column 0]: An import element must have at least one source

The usual cause of this error in my experience is when sources
are chosen using the 'arches' conditional. If the user specifies
an arch that the import element doesn't know about then they
will get see this error which may not be particularly helpful.