_pipeline.py: Better error handling from Pipeline.deps_elements()
Raise PipelineError in the case that specified --except elements
do not exist in the selected pipeline, instead of allowing an
action to continue when the user misspells an element in the --except
arguments.

Also make the default message for PipelineError an empty string,
so that intentionally unclassified errors just print as an empty string.