source.py: Ensure 'directory' exists in _stage
It is usually created by integration commands when staging dependencies,
but we cannot rely on this as some elements have no dependencies, and
some sources handle the non-existence of 'directory' better than others.

To avoid mandating that every source must be able to cope with
'directory's not existence, source will ensure it exists.