_sandboxbwrap.py: Fix job control in interactive mode
For some reason, launching bwrap under a new session (setsid()) causes
the shell launched in that container to lack job control.

This patch fixes this by conditionally creating the new session, which
we need to ensure we handle SIGTSTP and SIGTERM separately in regular
non-interactive mode, but not creating the session when launched in
interactive mode.

This fixes issue #40