_fuse/mount.py: Attempt to synchronize unmounts of fuse layers.
This is a defensive code path because of a BUG message which occured
with a stack trace, oddly causing os.makedirs() to report a FileExistsError
even with the exist_ok being set to True.

The directory in question was a fuse mountpoint we ensure the existence
of in _sandboxbwrap.py, and doing 'ls' on the directory shows the
error 'transport endpoint not connected'.

This indicates that the fuse mount still exists at this point but the
backing fuse process has already exited.