utils.py: Additional fix to _relative_symlink_target()
Since we merged MR 29 with a fix for resolving symlinks to realpaths,
we should also use the real path of the staging directory to produce
a relative path properly.

If we are considering that staged symlinks link to symlinked paths
inside the staging target, we must also consider that the staging
area itself can be in a symliked directory.