element.py: Removed _direct_deps() API
This API comes from a time when Element.dependencies() did not
have a `recurse` keyword argument (as we did not expect to require
that functionality to be public).

Since the same can be accomplished with Element.dependencies(... recurse=False),
it's better to just remove this old private API (which was actually broken and
recently fixed with commit b194179d422b65ce8eaa599901b0362840584adb).

NOTE: Tried this change on a pipeline with > 500 elements, and the
reported cache keys are the same using Element.dependencies() as they
were with the older Element._direct_deps() API.