utils.py: Fixed recent _set_deterministic_user() addition to not follow symlinks.
We want to change ownership of the symlinks, not their targets.

When doing this on an ostree checkout with symlinks to absolute paths, it
will make an attempt to modify ownership of files outside the sandbox,
and usually fail when building as the regular user (otherwise it will harm
the system when running as root).