_yaml.py: Added node_copy() and list_copy()
These are similar to node_chain_copy() and node_list_copy() which
are there mostly for optimization reasons, but node_copy() and
list_copy() entirely safely separate the nodes from their originals.

This is suitable when giving away a node to a plugin via an API
call where we want to be sure that no modifications modify the source
node.

As a side effect it also ensures that ever node has Provenance, in case
nodes are given back to the data model where they might be expected later.