_loader.py: Raise LoadError() for invalid variant requests.
  o When the toplevel variant does not exist in the loaded target,
    raise the LoadError

  o When any element requests a nonexistent variant on any of
    it's dependencies, raise the LoadError

  o Internal Dependency objects now track the provenance from
    whence they were declare, to allow more accurate error reporting
    for invalid variants via dependencies.

Beyond this, we also remove the Loader's self.loaded_files attribute
and a bogus assertion that two elements can claim the same name, instead
just rely on a single map of loaded elements to protect against reentrancy.

This was a legacy leftover from an initial implementation which allowed
elements to be named, and as such we were treating loaded files and
loaded elements separately.