docs: Fix Docker instructions
Suggest the user to mount /etc/passwd inside the container to fix `bzr`
command.

The command `bzr` fails if the uid of the user using it doesn't exist,
giving, for example, the following back trace when running `bzr init`:

      File "/usr/lib64/python2.7/site-packages/bzrlib/lockdir.py", line 238, in _attempt_lock
	tmpname = self._create_pending_dir()
      File "/usr/lib64/python2.7/site-packages/bzrlib/lockdir.py", line 335, in _create_pending_dir
	info = LockHeldInfo.for_this_process(self.extra_holder_info)
      File "/usr/lib64/python2.7/site-packages/bzrlib/lockdir.py", line 779, in for_this_process
	user=get_username_for_lock_info(),
      File "/usr/lib64/python2.7/site-packages/bzrlib/lockdir.py", line 863, in get_username_for_lock_info
	return osutils.getuser_unicode()
      File "/usr/lib64/python2.7/site-packages/bzrlib/osutils.py", line 356, in _posix_getuser_unicode
	name = getpass.getuser()
      File "/usr/lib64/python2.7/getpass.py", line 158, in getuser
	return pwd.getpwuid(os.getuid())[0]
    KeyError: 'getpwuid(): uid not found: 1000'

To fix this there were 2 possible solutions:
- To run everything as rooat (or as another user) in the container. We
will not need to set the uid to use, making the instructions simpler.
The main problem is that the files stored in the host will probably
have a different owner, and the user will have problems with file
permissions.
- To ensure the uid of the current user also exists inside the
container.