_frontend/main.py: Naming workspace functions with workspace prefix
The `list` function was already named `list_` to avoid clashing
with the builtin python type, the `open` command was causing log
viewing to fail because we try to call the `open` builtin python
function but it's shadowed.

So just call all of the workspace functions with a `workspace_`
prefix (but keep the same names on the command line).