_sandboxbwrap.py: Restore terminal after exit of interactive child
Make the main BuildStream process the foreground process again when the
interactive child exits. Otherwise the next read() on stdin will trigger
SIGTTIN and stop the process. This is required because the sandboxed
process does not have permission to do this on its own (running in
separate PID namespace).

dash still prints an error because it fails to restore the foreground
process, however, this is harmless. bash doesn't print an error in this
case, but the behavior is otherwise identical.

Fixes #41