workspace open: --no-checkout doesn't need --force
Remove the requirement to specify '--force' in conjunction with
'--no-checkout' if there are already files in the workspace. We won't
write anything when opening the workspace, so there's nothing to force.

For example, when opening a workspace to an existing clone of a
repository, it seems alarming to have to '--force' the workspace open.
It made me wonder if it will actually be overwritten.