_scheduler module: Create an event loop instead of using the default
Using the default thread context main loop means we reuse some of
the main context across invocations of buildstream in the same
process.

This is not an issue for buildstream itself but it impairs our
ability to run many test cases in a single process, so lets just
use a separate loop per Scheduler instance.