tests/sources/ostree.py: Removed http server test
This was failing intermittently if the test suite is
run multiple times (as it takes some time for the port 8000
to become available again).

Further, this test case was immaterial and not testing
any ostree functionality whatsoever. Presumably this was
a remnant from a time when the test case attempts to fetch
the payload over a local http connection.