artifactcache.py: Special casing local pull & push
It's interesting to keep the local directory push & pull
codepaths alive for the sake of testing everything that
calls into the artifact cache.

For this reason, even when the artifact cache is "offline",
it should be possible to pull and push to it if it is
a local artifact share.