Removing artifact cache tests.
Except for the ArtifactCache.remove() test, all of these assertions are
covered by the frontend `push` and `pull` tests. The remove() API itself
is unused except for being tested, lets leave this functionality out
until it's used.

The highlevel tests are easier to write and exercise more of the
codebase to assert the right behaviors, so lets only maintain that.