_frontend/main.py: Changes to the bst shell command
bst shell command is now an optional argument, which
defaults to ['sh', '-i']. This allows one to override
the command including the shell, so one need not invoke
the shell, thus it fixes issue #86.

Instead of an option. This means we can pick up a complete command
line after the '--' token directly from the remaining argument vector,
which makes escaping of things a bit less complicated.

E.g.:

  bst shell --scope run <target.bst> -- sh -c "echo Hello World"

Further, the --scope argument is removed in favor of defaulting
to the --scope=run behavior, and adding a --build option
(use `bst shell --build <target>` for build shells, otherwise
juse use `bst shell <target>` for run shells).

Finally, the --builddir has been renamed to --sysroot; which
a bit more accurately describes what it is.