Let users override project artifacts options
A project can specify its artifacts cache sharing settings, and users can
define a default artifacts cache to use as a fallback.

With this change, users can also override the project configuration with
their own.

That means for a project named "libfoo", BuildStream will resolve the
artifacts-related options in the following order:

1. the projects.libfoo.artifacts options from the user configuration;
2. if the above was not defined, then the artifacts options from the
   project configuration;
3. if the above was not defined, then the artifacts options from the
   user configuration;

Fixes #87