fuse.py: Add support for Linux on ppc64
This is equivalent to this upstream pull request:

  https://github.com/terencehonles/fusepy/pull/72

It is needed for BuildStream to work on ppc64 architectures. Without this
fix, the FUSE safe-hardlinks filesystem returns garbage. One visible effect
is that running integration commands fails with "executable format error"
(due to the filesystem returning garbage in place of the actual binary).

https://gitlab.com/BuildStream/buildstream/issues/83