Fix timestamps in Python bytecode
When building Python modules, a bytecode `.pyc` file is generated from
the source `.py` file.

The former contains 4 bytes representing the timestamp of the latter at
the time it was generated.

Unfortunately, after building OSTree sets all the file timestamps to 0,
which introduces a discrepency between the timestamp of the `.py` file
and the 4 bytes stored inside the `.pyc` file.

As a result, when running a Python module from a checkout, Python thinks
the bytecode files are stale, which causes a dramatic performance
penalty when starting an application.

Fixes #94