.gitlab-ci.yml: Follow latest BuildStream Docker image
This allows us to remove env vars that are now set in the image,
and some packages which are already installed in the image.

I've left Sphinx (which is needed for the documentation task) out of the
image as it pulls in a bunch of Python 2 dependencies that would bloat
the image. Although note that Python 2 is already present in the image
because Bazaar depends on it.