_yaml.py: Implement array composition directives
Now we support (<), (>) and (=) directives for declarative list
composition. In place of specifying a list directly to override
another list, specify a dict with sublists for prepending, appending
and overriding, respectively.

As a consequence, the composite policy which previously decided
between whether to override or append has been removed from the API.

Further, the typesafe parameter has been removed and all
compositions are expected to be typesafe (this was always being
passed as True anyway).

This also adds _yaml.node_final_assertions() which needs to be
called after the final composition of a dictionary, before
validation and before trying to parse values from the said node.

This allows the _yaml module to have a chance to make some
assertions after everything is composited