_yaml.py: Fixes #119 - Custom implementation of ChainMap
We had previously painted ourselves into a corner by using ChainMap
as an optimization for creating mutable copies of a loaded YAML tree.

The problem with this is that, deleting things which have not been
explicitly mutated raises a KeyError, because ChainMap only operates
on the first (mutable) mapping; and also fails to preserve explicitly
deleted state.

This patch adds a custom derived ChainMap which caches deleted state
in self.__deletions and behaves more like a copy-on-write mutable
dictionary.