tests/sources/generic.py: Removing test, no tests should ever be added here
New tests only via the frontend cli, never mucking about with source APIs
directly, this whole directory needs to eventually be removed.

This test was also testing the wrong condition, it's okay that a source
stage to a non-empty directory, it's not okay that an element stages
it's group of sources to a non-empty directory.