data/projectconfig.yaml: Issue #106 - Compress debug symbols by default
Deconstructed the default strip-commands to declare some objcopy and
strip arguments separately, so it's easier to override these in the
case that you are building with an older toolchain which does not
support compression of debug symbols (or if you want to strip them
with something in a more exotic way).

By default pass --compress-debug-sections to objcopy.