Add support for doing incremental builds
This functionality is only supported for sources which have an open
workspace. When such sources are present, the workspace directory will
be mounted directly inside the sandbox. As opposed to the default
behavior, which is to copy files inside the sandbox.

This will save time when building large projects as only those files
will need be re-compiled that have been modified during two consecutive
builds (assuming the underlying build system supports such behavior).

A few things to note regarding this behavior:

- If there are any `configure-commands` present, they will run only once
  for each open workspace. If an element has multiple workspaces and any
  one of them is opened/closed, they will be executed again on the next
  run. But, modifying the contents of a workspace will not trigger the
  `configure-commands` to be executed on the next run.

- Workspaced builds still leverage the cache. So, if no changes are made
  to the workspace, i.e. no files are modified, then it will not force a
  rebuild.

Fixes #192.