_pipeline.py: The toplevel message() function no longer takes an element.
Instead, use None as the unique ID.

This will help with messaging when we allow invoking multiple targets
and the element would have been ambiguous - this also consequently
fixes issue #137.

The reason for the hangs with #137 is because:
 o When you --except a base element, reverse dependencies cannot calculate a cache key
 o When you track, cache keys are intentionally reset at startup time, because
   we know they are going to change
 o At the end of tracking, we make one attempt to print the toplevel cache key

This operation is insanely expensive, because we never cache a cache key because
it logically cannot ever be resolved in this situation.

This fix is basically a workaround to the above.