_platform/linux.py: Add preflight check to detect user namespaces
Here we check if `bwrap --ro-bind / / --unshare-user --uid 0 --gid 0 whoami`
returns successfully and prints 'root', to check if we are indeed capable
of creating user namespaces on the host.

If we are unable, then we save that state in the platform allowing
buildstream to behave differently, and print an informative startup
warning about this.

This fixes issue #92