plugins/sources/tar.py: Ignore possible leading '.' directory.
To extract the full tarball, one should set base-dir to an
empty string.

By ignoring the leading '.' in any archive, we make the 'base-dir'
API more predictable and reliable - the default behavior of '*' is
to pickup the first directory in the tarball (usually source code
tarballs are encoded with one leading directory) - in the off chance
that a source tarball has a leading '.' in it; that would cause
the 'base-dir' default '*' glob to extract the whole thing.

It seems undesirable to behave differently depending on whether
a tarball was encoded with, or without a leading '.'