Log details of artifact splitting when building 'compose' elements
Artifact split rules are complicated to reason about. There needs to be
some way of seeing why a 'compose' artifact included a given file.

The log file now contains a long message like this, listing every
artifact, domain and file:

    [--:--:--] INFO    [initramfs/initramfs.bst]: Composed 2369 files

        Integration
          - var/cache/ldconfig/aux-cache

        gnu-toolchain/binutils.bst runtime
          - usr/x86_64-unknown-linux-gnu
          - usr/x86_64-unknown-linux-gnu/bin/ar
          - usr/x86_64-unknown-linux-gnu/bin/as
          - usr/x86_64-unknown-linux-gnu/bin/ld

        gnu-toolchain/make.bst (not in any domain)
          - usr
          - usr/lib
          - usr/share

        ...

Note that a file can be in multiple domains.

Size of the log files is a concern and it may be that in future we
'compress' some of these entries, e.g. if all files in a given
directory come from one artifact then we just need to list the
directory, not every filename.