autotools: Always run the autogen step
BuildStream made the choice to unconditionally run the autogen step for
autotools elements, even when building from tarballs.

However, due to how the conditionals for what to do in that step were
set up, some tarballs would fall through the crack and effectively skip
that step.

For example the Python tarballs and VCS checkouts both:

* do not have an autogen/autogen.sh script
* do not have a bootstrap/bootstrap.sh script
* do have a configure script

The conditionals we had before this change would not do anything for a
project like that, and the first thing to actually be run would be the
configure script.

With this change, `autoreconf` gets run whether the configure script
exists or not, ensuring we always effectively do the autogen step.

More details on the mailing-list:

    https://mail.gnome.org/archives/buildstream-list/2017-November/msg00015.html